+++ Projects/Android/jni/rtcw/src/game/g_local.h
@@ -1218,7 +1218,7 @@ void    trap_GetUsercmd( int clientNum, usercmd_t *cmd );
 qboolean    trap_GetEntityToken( char *buffer, int bufferSize );
 qboolean trap_GetTag( int clientNum, char *tagName, orientation_t * or );
 
-int trap_Vibrate(float duration, int channel, float intensity );
+int trap_Vibrate(int duration, int channel, float intensity );
 
 int     trap_DebugPolygonCreate( int color, int numPoints, vec3_t *points );
 void    trap_DebugPolygonDelete( int id );
